Rename udf to matchbox, and remove deprecated keep/discard Python implementations.
