+++ src/main/python/aut/__init__.py
@@ -1,6 +1,6 @@
 from aut.common import WebArchive
-from aut.filters import discard_mime_types, filter_mime_types, keep_valid_pages
-from aut.udfs import (
+from aut.filters import keep_valid_pages
+from aut.matchbox import (
     compute_MD5,
     compute_SHA1,
     extract_domain,
@@ -14,8 +14,6 @@ __all__ = [
     "compute_MD5",
     "compute_SHA1",
     "extract_domain",
-    "discard_mime_types",
-    "filter_mime_types",
     "keep_valid_pages",
     "remove_html",
     "remove_prefix_www",
