+++ src/main/python/aut/filters.py
@@ -23,17 +23,3 @@ def keep_valid_pages(df):
         )
         .filter(col("http_status_code") == 200)
     )
-
-
-def filter_mime_types(mime_types):
-    return df.filter(filtered_mime_type(col("mime_type_web_server")))
-
-
-filter_mime_types = udf(filter_mime_types.contains(mime_types), ArrayType())
-
-
-def discard_mime_types(mime_types):
-    return df.filter(filtered_mime_type(col("mime_type_web_server")))
-
-
-discard_mime_types = udf(~discard_mime_types.contains(mime_types), ArrayType())
